requirements.txt
# FastAPI and web server
fastapi
uvicorn[standard]
requests

# For authentication
python-jose[cryptography]
passlib

# For database
sqlalchemy

# PDF and document parsing
PyMuPDF       # For PDF parsing
markdown      # For .md to HTML conversion
beautifulsoup4 # For HTML tag stripping

# ChromaDB + embeddings
chromadb
sentence-transformers
tiktoken
markdown

nltk

# For LLMs
openai

# Pydantic for schema validation
pydantic

# Utility libraries
numpy
scipy

# For file uploads
python-multipart
